Refactor the high-level client
It contains a type-erased service which is built from components that can add features to the client
