+++ Cargo.toml
@@ -62,6 +62,12 @@ version = "0.4"
 features = ["make", "util"]
 default-features = false
 
+[dependencies.tower-http]
+version = "0.5.2"
+features = ["follow-redirect", "set-header", "trace"]
+default-features = false
+optional = true
+
 [dev-dependencies]
 clap = { version = "4.5.7" }
 color-eyre = "0.6"
@@ -73,7 +79,7 @@ webpki-roots.version = "0.26"
 
 [features]
 axum = ["dep:axum"]
-client = ["incoming", "dep:socket2", "dep:thiserror"]
+client = ["incoming", "dep:socket2", "dep:thiserror", "dep:tower-http"]
 default = ["client", "server", "discovery", "stream"]
 discovery = ["server", "client", "pidfile", "stream", "dep:dashmap"]
 docs = []
