+++ src/client/builder.rs
@@ -1,56 +1,134 @@
 #[cfg(feature = "tls")]
 use std::sync::Arc;
 
+use http::HeaderValue;
 #[cfg(feature = "tls")]
 use rustls::ClientConfig;
+use tower::ServiceBuilder;
+use tower_http::follow_redirect::FollowRedirectLayer;
+use tower_http::set_header::SetRequestHeaderLayer;
 
-use super::conn::transport::TransportExt as _;
+use super::conn::protocol::auto;
+use super::conn::transport::tcp::TcpTransportConfig;
+use super::conn::transport::TransportExt;
+use super::conn::Connection;
+use super::conn::Protocol;
+use super::conn::Transport;
+use super::pool::PoolableConnection;
 use super::ClientService;
-use super::ClientTlsTcpService;
+use crate::client::conn::connection::ConnectionError;
 #[cfg(feature = "tls")]
 use crate::client::default_tls_config;
 use crate::client::{conn::protocol::auto::HttpConnectionBuilder, Client};
+use crate::info::HasConnectionInfo;
+use crate::service::SharedService;
+
+pub trait BuildProtocol<IO>
+where
+    IO: HasConnectionInfo,
+{
+    type Target: Protocol<IO>;
+    fn build(self) -> Self::Target;
+}
+
+impl<P, IO> BuildProtocol<IO> for P
+where
+    P: Protocol<IO>,
+    IO: HasConnectionInfo,
+{
+    type Target = P;
+    fn build(self) -> Self::Target {
+        self
+    }
+}
+
+pub trait BuildTransport {
+    type Target: Transport;
+    fn build(self) -> Self::Target;
+}
+
+impl<T> BuildTransport for T
+where
+    T: Transport,
+{
+    type Target = T;
+    fn build(self) -> Self::Target {
+        self
+    }
+}
 
 /// A builder for a client.
 #[derive(Debug)]
-pub struct Builder {
-    tcp: crate::client::conn::transport::tcp::TcpTransportConfig,
+pub struct Builder<T, P> {
+    transport: T,
+    protocol: P,
+    user_agent: Option<String>,
     #[cfg(feature = "tls")]
     tls: Option<ClientConfig>,
     pool: Option<crate::client::pool::Config>,
-    conn: HttpConnectionBuilder,
 }
 
-impl Default for Builder {
-    fn default() -> Self {
+impl Builder<(), ()> {
+    /// Create a new, empty builder
+    pub fn new() -> Self {
         Self {
-            tcp: Default::default(),
+            transport: (),
+            protocol: (),
+            user_agent: None,
             #[cfg(feature = "tls")]
-            tls: Some(default_tls_config()),
-            pool: Some(Default::default()),
-            conn: Default::default(),
+            tls: None,
+            pool: None,
         }
     }
 }
 
-impl Builder {
+impl Default for Builder<TcpTransportConfig, HttpConnectionBuilder> {
+    fn default() -> Self {
+        Self {
+            transport: Default::default(),
+            protocol: Default::default(),
+            user_agent: None,
+            #[cfg(feature = "tls")]
+            tls: Some(default_tls_config()),
+            pool: Some(Default::default()),
+        }
+    }
+}
+
+impl<T, P> Builder<T, P> {
     /// Use the provided TCP configuration.
-    pub fn with_tcp(
-        mut self,
-        config: crate::client::conn::transport::tcp::TcpTransportConfig,
-    ) -> Self {
-        self.tcp = config;
-        self
+    pub fn with_tcp(self, config: TcpTransportConfig) -> Builder<TcpTransportConfig, P> {
+        Builder {
+            transport: config,
+            protocol: self.protocol,
+            user_agent: self.user_agent,
+            #[cfg(feature = "tls")]
+            tls: self.tls,
+            pool: self.pool,
+        }
     }
 
-    /// TCP configuration.
-    pub fn tcp(&mut self) -> &mut crate::client::conn::transport::tcp::TcpTransportConfig {
-        &mut self.tcp
+    /// Provide a custom transport
+    pub fn with_transport<T2>(self, transport: T2) -> Builder<T2, P> {
+        Builder {
+            transport,
+            protocol: self.protocol,
+            user_agent: self.user_agent,
+            #[cfg(feature = "tls")]
+            tls: self.tls,
+            pool: self.pool,
+        }
     }
 }
 
 #[cfg(feature = "tls")]
-impl Builder {
+impl<T, P> Builder<T, P> {
+    /// Disable TLS
+    pub fn without_tls(mut self) -> Self {
+        self.tls = None;
+        self
+    }
+
     /// Use the provided TLS configuration.
     pub fn with_tls(mut self, config: ClientConfig) -> Self {
         self.tls = Some(config);
@@ -69,15 +147,29 @@ impl Builder {
     }
 }
 
-impl Builder {
+#[cfg(not(feature = "tls"))]
+impl<T, P> Builder<T, P> {
+    /// Disable TLS
+    pub fn without_tls(self) -> Self {
+        self
+    }
+}
+
+impl<T, P> Builder<T, P> {
     /// Connection pool configuration.
-    pub fn pool(&mut self) -> &mut Option<crate::client::pool::Config> {
-        &mut self.pool
+    pub fn pool(&mut self) -> Option<&mut crate::client::pool::Config> {
+        self.pool.as_mut()
     }
 
     /// Use the provided connection pool configuration.
     pub fn with_pool(mut self, pool: crate::client::pool::Config) -> Self {
         self.pool = Some(pool);
+        self
+    }
+
+    /// Configure the default pool settings
+    pub fn with_default_pool(mut self) -> Self {
+        self.pool = Some(Default::default());
         self
     }
 
@@ -88,86 +180,112 @@ impl Builder {
     }
 }
 
-impl Builder {
+impl<T, P> Builder<T, P> {
+    /// Use the auto-HTTP Protocol
+    pub fn with_auto_http(self) -> Builder<T, auto::HttpConnectionBuilder> {
+        Builder {
+            transport: self.transport,
+            protocol: auto::HttpConnectionBuilder::default(),
+            user_agent: self.user_agent,
+            #[cfg(feature = "tls")]
+            tls: self.tls,
+            pool: self.pool,
+        }
+    }
+
     /// Use the provided HTTP connection configuration.
-    pub fn with_conn(
-        mut self,
-        conn: crate::client::conn::protocol::auto::HttpConnectionBuilder,
-    ) -> Self {
-        self.conn = conn;
-        self
+    pub fn with_protocol<P2>(self, protocol: P2) -> Builder<T, P2> {
+        Builder {
+            transport: self.transport,
+            protocol,
+            user_agent: self.user_agent,
+            #[cfg(feature = "tls")]
+            tls: self.tls,
+            pool: self.pool,
+        }
     }
 
     /// HTTP connection configuration.
-    pub fn conn(&mut self) -> &mut crate::client::conn::protocol::auto::HttpConnectionBuilder {
-        &mut self.conn
+    pub fn protocol(&mut self) -> &mut P {
+        &mut self.protocol
     }
 }
 
-impl Builder {
+impl<T, P> Builder<T, P>
+where
+    T: BuildTransport,
+    <T as BuildTransport>::Target: Transport + Clone + Send + Sync + 'static,
+    <<T as BuildTransport>::Target as Transport>::IO:
+        tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin,
+    <<<T as BuildTransport>::Target as Transport>::IO as HasConnectionInfo>::Addr:
+        Unpin + Clone + Send,
+    P: BuildProtocol<super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>>,
+    <P as BuildProtocol<
+        super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>,
+    >>::Target: Protocol<
+            super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>,
+            Error = ConnectionError,
+        > + Clone
+        + Send
+        + Sync
+        + 'static,
+    <<P as BuildProtocol<
+        super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>,
+    >>::Target as Protocol<
+        super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>,
+    >>::Connection: PoolableConnection,
+    crate::Body: From<
+        <<<P as BuildProtocol<
+            super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>,
+        >>::Target as Protocol<
+            super::conn::stream::Stream<<<T as BuildTransport>::Target as Transport>::IO>,
+        >>::Connection as Connection>::ResBody,
+    >,
+{
     /// Build the client.
-    pub fn build(self) -> Client<ClientTlsTcpService> {
-        Client {
-            service: ClientService {
-                #[cfg(feature = "tls")]
-                transport: crate::client::conn::transport::tcp::TcpTransport::builder()
-                    .with_config(self.tcp)
-                    .with_gai_resolver()
-                    .build()
-                    .with_optional_tls(self.tls.map(Arc::new)),
+    pub fn build(self) -> Client {
+        let user_agent = if let Some(ua) = self.user_agent {
+            HeaderValue::from_str(&ua).expect("user-agent should be a valid http header")
+        } else {
+            HeaderValue::from_static(concat!(
+                env!("CARGO_PKG_NAME"),
+                "/",
+                env!("CARGO_PKG_VERSION")
+            ))
+        };
 
-                #[cfg(not(feature = "tls"))]
-                transport: crate::client::conn::transport::tcp::TcpTransport::builder()
-                    .with_config(self.tcp)
-                    .with_gai_resolver()
-                    .build()
-                    .without_tls(),
+        #[cfg(feature = "tls")]
+        let transport = self
+            .transport
+            .build()
+            .with_optional_tls(self.tls.map(Arc::new));
+        #[cfg(not(feature = "tls"))]
+        let transport = self.transport.build().without_tls();
 
-                protocol: HttpConnectionBuilder::default(),
-                pool: self.pool.map(crate::client::pool::Pool::new),
-
+        let service = ServiceBuilder::new()
+            .layer(SharedService::layer())
+            .layer(SetRequestHeaderLayer::if_not_present(
+                http::header::USER_AGENT,
+                user_agent,
+            ))
+            .layer(FollowRedirectLayer::new())
+            .service(ClientService {
+                transport,
+                protocol: self.protocol.build(),
+                pool: self.pool.map(super::pool::Pool::new),
                 _body: std::marker::PhantomData,
-            },
-        }
+            });
+
+        Client::new_from_service(service)
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use super::Builder;
 
     #[test]
-    fn test_builder() {
-        let client = Builder::default().build();
-        assert!(client.service.pool.is_some());
-    }
-
-    #[test]
-    fn test_builder_tcp() {
-        let mut builder = Builder::default();
-        builder.tcp().nodelay = true;
-
-        let client = builder.build();
-        assert!(client.service.transport.inner().config().nodelay)
-    }
-
-    #[cfg(feature = "tls")]
-    #[test]
-    fn test_builder_tls() {
-        let mut builder = Builder::default();
-        let mut tls = super::default_tls_config();
-        tls.alpn_protocols.push(b"a1".to_vec());
-        builder = builder.with_tls(tls);
-
-        let client = builder.build();
-        assert_eq!(
-            client
-                .service
-                .transport
-                .tls_config()
-                .unwrap()
-                .alpn_protocols,
-            vec![b"h2".to_vec(), b"http/1.1".to_vec(), b"a1".to_vec()]
-        );
+    fn build_default_compiles() {
+        let _ = Builder::default().build();
     }
 }
