+++ src/client/conn/protocol/mod.rs
@@ -10,6 +10,7 @@ use futures_core::future::BoxFuture;
 use tokio::io::AsyncRead;
 use tokio::io::AsyncWrite;
 use tower::Service;
+use tracing::Instrument;
 
 use super::connection::ConnectionError;
 use super::connection::HttpConnection;
@@ -168,20 +169,27 @@ where
         let builder = self.clone();
         let stream = req.transport.into_inner();
 
+        let info = stream.info();
+        let span = tracing::info_span!("connection", version=?http::Version::HTTP_11, peer=%info.remote_addr());
+
         Box::pin(async move {
             let (sender, conn) = builder
                 .handshake(TokioIo::new(stream))
                 .await
                 .map_err(|err| ConnectionError::Handshake(err.into()))?;
-            tokio::spawn(async {
-                if let Err(err) = conn.await {
-                    if err.is_user() {
-                        tracing::error!(err = format!("{err:#}"), "h1 connection driver error");
-                    } else {
-                        tracing::debug!(err = format!("{err:#}"), "h1 connection driver error");
+
+            tokio::spawn(
+                async {
+                    if let Err(err) = conn.await {
+                        if err.is_user() {
+                            tracing::error!(err = format!("{err:#}"), "h1 connection driver error");
+                        } else {
+                            tracing::debug!(err = format!("{err:#}"), "h1 connection driver error");
+                        }
                     }
                 }
-            });
+                .instrument(span),
+            );
             Ok(HttpConnection::h1(sender))
         })
     }
@@ -212,21 +220,26 @@ where
     fn call(&mut self, req: ProtocolRequest<IO>) -> Self::Future {
         let builder = self.clone();
         let stream = req.transport.into_inner();
+        let info = stream.info();
+        let span = tracing::info_span!("connection", version=?http::Version::HTTP_11, peer=%info.remote_addr());
 
         Box::pin(async move {
             let (sender, conn) = builder
                 .handshake(TokioIo::new(stream))
                 .await
                 .map_err(|err| ConnectionError::Handshake(err.into()))?;
-            tokio::spawn(async {
-                if let Err(err) = conn.await {
-                    if err.is_user() {
-                        tracing::error!(err = format!("{err:#}"), "h2 connection driver error");
-                    } else {
-                        tracing::debug!(err = format!("{err:#}"), "h2 connection driver error");
+            tokio::spawn(
+                async {
+                    if let Err(err) = conn.await {
+                        if err.is_user() {
+                            tracing::error!(err = format!("{err:#}"), "h2 connection driver error");
+                        } else {
+                            tracing::debug!(err = format!("{err:#}"), "h2 connection driver error");
+                        }
                     }
                 }
-            });
+                .instrument(span),
+            );
             Ok(HttpConnection::h2(sender))
         })
     }
