+++ src/service/mod.rs
@@ -6,6 +6,7 @@ mod http;
 mod incoming;
 mod make;
 mod serviceref;
+mod shared;
 
 pub use self::adapt::{AdaptCustomBodyExt, AdaptCustomBodyLayer, AdaptCustomBodyService};
 pub use self::adapt::{AdaptOuterBodyLayer, AdaptOuterBodyService};
@@ -14,4 +15,5 @@ pub use self::http::HttpService;
 pub use self::incoming::{AdaptIncomingLayer, AdaptIncomingService};
 pub use self::make::{make_service_fn, BoxMakeServiceLayer, BoxMakeServiceRef, MakeServiceRef};
 pub use serviceref::ServiceRef;
+pub use shared::SharedService;
 pub use tower::{service_fn, Service, ServiceBuilder, ServiceExt};
