+++ tests/httpbin.rs
@@ -2,24 +2,20 @@ use http_body_util::BodyExt as _;
 use hyperdriver::client::{conn::transport::tcp::TcpTransportConfig, Client};
 
 /// Make a request to httpbin.org
+#[tracing::instrument(skip_all, fields(method = %req.method(), version = ?req.version()))]
 async fn httpbin_request(
-    mut req: http::Request<hyperdriver::Body>,
+    req: http::Request<hyperdriver::Body>,
 ) -> Result<bytes::Bytes, Box<dyn std::error::Error + Send + Sync + 'static>> {
     let config = TcpTransportConfig {
         happy_eyeballs_timeout: Some(std::time::Duration::from_secs(1)),
         ..Default::default()
     };
 
-    let mut client = Client::builder()
+    let client = Client::builder()
         .with_tcp(config)
+        .with_auto_http()
         .with_default_tls()
         .build();
-
-    let hdrs = req.headers_mut();
-    hdrs.append(
-        http::header::USER_AGENT,
-        concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION")).parse()?,
-    );
 
     let res = client.request(req).await?;
 
@@ -36,6 +32,9 @@ fn request(method: http::Method, version: http::Version) -> http::request::Build
 
 #[tokio::test]
 async fn get_h1() {
+    let _ = tracing_subscriber::fmt::try_init();
+    let _ = rustls::crypto::ring::default_provider().install_default();
+
     let req = request(http::Method::GET, http::Version::HTTP_11)
         .body(hyperdriver::Body::empty())
         .unwrap();
@@ -45,6 +44,9 @@ async fn get_h1() {
 
 #[tokio::test]
 async fn get_h2() {
+    let _ = tracing_subscriber::fmt::try_init();
+    let _ = rustls::crypto::ring::default_provider().install_default();
+
     let req = request(http::Method::GET, http::Version::HTTP_2)
         .body(hyperdriver::Body::empty())
         .unwrap();
@@ -54,6 +56,9 @@ async fn get_h2() {
 
 #[tokio::test]
 async fn post_h1() {
+    let _ = tracing_subscriber::fmt::try_init();
+    let _ = rustls::crypto::ring::default_provider().install_default();
+
     let req = request(http::Method::POST, http::Version::HTTP_11)
         .body(hyperdriver::Body::from("Hello, world!"))
         .unwrap();
@@ -63,6 +68,9 @@ async fn post_h1() {
 
 #[tokio::test]
 async fn post_h2() {
+    let _ = tracing_subscriber::fmt::try_init();
+    let _ = rustls::crypto::ring::default_provider().install_default();
+
     let req = request(http::Method::POST, http::Version::HTTP_2)
         .body(hyperdriver::Body::from("Hello, world!"))
         .unwrap();
