+++ tests/tls.rs
@@ -105,11 +105,12 @@ async fn tls_echo_h1() {
         .with_no_client_auth();
     client_tls.alpn_protocols.push(b"http/1.1".to_vec());
 
-    let mut client = Client::new(
-        Builder::new(),
-        DuplexTransport::new(1024, duplex_client).with_tls(client_tls.into()),
-        Default::default(),
-    );
+    let client = Client::builder()
+        .with_protocol(Builder::new())
+        .with_default_pool()
+        .with_transport(DuplexTransport::new(1024, duplex_client))
+        .with_tls(client_tls)
+        .build();
 
     let response: Response<hyperdriver::Body> = client
         .request(
@@ -154,11 +155,10 @@ async fn tls_echo_h2() {
         .with_no_client_auth();
     client_tls.alpn_protocols.push(b"h2".to_vec());
 
-    let mut client = Client::new(
-        Builder::new(TokioExecutor::new()),
-        DuplexTransport::new(1024, duplex_client).with_tls(client_tls.into()),
-        Default::default(),
-    );
+    let client = Client::builder()
+        .with_protocol(Builder::new(TokioExecutor::new()))
+        .with_transport(DuplexTransport::new(1024, duplex_client).with_tls(client_tls.into()))
+        .build();
 
     let response: Response<hyperdriver::Body> = client
         .request(
